move biometric_service declaration to /services and method channel implementation to /method_channel
