+++ lib/gigya_flutter_plugin.dart
@@ -1,7 +1,7 @@
 import 'src/models/enums/social_provider.dart';
 import 'src/models/screenset_event.dart';
 import 'src/platform_interface/gigya_flutter_plugin_platform_interface.dart';
-import 'src/services/biometric_service/biometric_service.dart';
+import 'src/services/biometric_service.dart';
 import 'src/services/interruption_resolver.dart';
 import 'src/services/otp_service.dart';
 import 'src/services/web_authentication_service.dart';
@@ -27,7 +27,7 @@ export 'src/models/screenset_event.dart';
 export 'src/models/session_info.dart';
 export 'src/models/skill.dart';
 export 'src/models/work.dart';
-export 'src/services/biometric_service/biometric_service.dart';
+export 'src/services/biometric_service.dart';
 export 'src/services/interruption_resolver.dart';
 export 'src/services/otp_service.dart';
 export 'src/services/web_authentication_service.dart';
