+++ lib/src/method_channel/gigya_flutter_plugin_method_channel.dart
@@ -8,11 +8,11 @@ import '../models/enums/social_provider.dart';
 import '../models/gigya_error.dart';
 import '../models/screenset_event.dart';
 import '../platform_interface/gigya_flutter_plugin_platform_interface.dart';
-import '../services/biometric_service/biometric_service.dart';
-import '../services/biometric_service/method_channel_biometric_service.dart';
+import '../services/biometric_service.dart';
 import '../services/interruption_resolver.dart';
 import '../services/otp_service.dart';
 import '../services/web_authentication_service.dart';
+import 'method_channel_biometric_service.dart';
 import 'method_channel_interruption_resolver.dart';
 import 'method_channel_otp_service.dart';
 import 'method_channel_web_authentication_service.dart';
