+++ lib/src/method_channel/method_channel_biometric_service.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/services.dart' show MethodChannel, PlatformException;
 import 'package:gigya_flutter_plugin/gigya_flutter_plugin.dart';
 
-import '../../models/enums/methods.dart';
+import '../models/enums/methods.dart';
 
 /// This class represents a [BiometricService] that uses a [MethodChannel]
 /// for its implementation.
