+++ lib/src/services/biometric_service.dart
